+++ forge/src/main/java/io/github/flemmli97/advancedgolems/forge/AdvancedGolemsForge.java
@@ -2,14 +2,17 @@ package io.github.flemmli97.advancedgolems.forge;
 
 import io.github.flemmli97.advancedgolems.AdvancedGolems;
 import io.github.flemmli97.advancedgolems.entity.GolemBase;
+import io.github.flemmli97.advancedgolems.events.EventCalls;
 import io.github.flemmli97.advancedgolems.forge.client.ClientInit;
 import io.github.flemmli97.advancedgolems.forge.config.ConfigLoader;
 import io.github.flemmli97.advancedgolems.forge.config.ConfigSpecs;
 import io.github.flemmli97.advancedgolems.registry.ModEntities;
 import io.github.flemmli97.advancedgolems.registry.ModItems;
+import net.minecraft.world.phys.EntityHitResult;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.entity.EntityAttributeCreationEvent;
+import net.minecraftforge.event.entity.ProjectileImpactEvent;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
@@ -33,6 +36,7 @@ public class AdvancedGolemsForge {
         }
         modbus.addListener(AdvancedGolemsForge::configReload);
         modbus.addListener(AdvancedGolemsForge::attributeEvent);
+        MinecraftForge.EVENT_BUS.addListener(AdvancedGolemsForge::projectileEvent);
     }
 
     public static void attributeEvent(EntityAttributeCreationEvent event) {
@@ -42,5 +46,10 @@ public class AdvancedGolemsForge {
     public static void configReload(ModConfigEvent event) {
         if (event.getConfig().getSpec() == ConfigSpecs.commonSpec)
             ConfigLoader.load();
+    }
+
+    public static void projectileEvent(ProjectileImpactEvent event) {
+        if (event.getRayTraceResult() instanceof EntityHitResult entityHitResult && !EventCalls.canProjectileHit(event.getProjectile(), entityHitResult))
+            event.setCanceled(true);
     }
 }